bugs -_-

+++ src/database.py
@@ -3,6 +3,7 @@ import sqlite3
 from sqlite3 import Error
 from users import roles
 from cryptoUtils import cryptoUtils
+import os
 
 
 class DB:
@@ -16,13 +17,13 @@ class DB:
             first_name TEXT NOT NULL,
             last_name TEXT NOT NULL,
             age INTEGER CHECK(age >= 0),
-            gender TEXT CHECK(gender IN ('Male', 'Female', 'Other')),
+            gender TEXT NOT NULL,
             weight REAL CHECK(weight >= 0),
-            address TEXT,
-            city TEXT CHECK(city IN ('Amsterdam', 'Rotterdam', 'The Hague', 'Utrecht', 'Eindhoven', 'Tilburg', 'Groningen', 'Almere', 'Breda', 'Nijmegen')),
+            address TEXT NOT NULL,
+            city TEXT,
             postalCode TEXT,
-            email TEXT UNIQUE NOT NULL,
-            mobile TEXT UNIQUE NOT NULL,
+            email TEXT,
+            mobile TEXT,
             registration_date DATE NOT NULL
         )
         """
@@ -44,10 +45,10 @@ class DB:
             first_name TEXT NOT NULL,
             last_name TEXT NOT NULL,
             username TEXT UNIQUE NOT NULL,
-            password TEXT NOT NULL,
+            password_hash TEXT NOT NULL,
             registration_date DATE NOT NULL,
-            role TEXT CHECK(role IN ('admin', 'consultant', 'superadmin')) NOT NULL,
-            temp BOOLEAN NOT NULL CHECK(temp IN (0, 1)),
+            role TEXT NOT NULL,
+            temp BOOLEAN NOT NULL,
             salt TEXT NOT NULL
         )
         """
@@ -67,33 +68,42 @@ class DB:
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
+            public_key = cryptoUtils.loadPublicKey()
+            private_key = cryptoUtils.loadPrivateKey()
 
-            cursor.execute("SELECT COUNT(*) FROM users WHERE username = 'super_admin'")
-            user_exists = cursor.fetchone()[0] > 0
+            cursor.execute("SELECT * FROM users")
+            users = cursor.fetchall()
 
-            if not user_exists:
+            superadmin_exists = False
+
+            for user in users:
+                decrypted_role = cryptoUtils.decryptWithPrivateKey(private_key, user[6])  # Assuming role is stored as encrypted bytes in user[7]
+                if decrypted_role == b"superadmin":
+                    superadmin_exists = True
+                    break
+
+            if superadmin_exists:
+                print("Superadmin already exists.")
+            else:
                 hashed_password, salt = cryptoUtils.hashPassword("Admin_123?")
-                
+                encrypted_username = cryptoUtils.encryptWithPublicKey(public_key, "super_admin")
+                encrypted_role = cryptoUtils.encryptWithPublicKey(public_key, "superadmin")
+
                 query = """
-                INSERT INTO users (first_name, last_name, username, password, registration_date, role, temp,salt)
+                INSERT INTO users (first_name, last_name, username, password_hash, registration_date, role, temp, salt)
                 VALUES (?, ?, ?, ?, ?, ?, ?, ?)
                 """
                 registration_date = date.today().strftime("%Y-%m-%d")
-                parameters = ("Kawish", "Gangadin", "super_admin", hashed_password, registration_date, "superadmin", 0, salt)
+                parameters = ("Kawish", "Gangadin", encrypted_username, hashed_password, registration_date, encrypted_role, 0, salt)
 
                 cursor.execute(query, parameters)
                 conn.commit()
                 print("Superadmin initialized successfully.")
-            else:
-                print("Superadmin already exists.")
-
         except sqlite3.Error as e:
             print("An error occurred while initializing superadmin:", e)
-
         finally:
             if conn:
                 conn.close()
-
 
     
     def searchMember(self, search_key):
@@ -130,11 +140,24 @@ class DB:
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
-            query = "SELECT * FROM users WHERE username = ?"
-            cursor.execute(query, (username,))
-            user = cursor.fetchone()
+            query = "SELECT * FROM users"
+            cursor.execute(query)
+            users = cursor.fetchall()
             cursor.close()
-            return user
+
+            # Decrypt and find the user
+            private_key = cryptoUtils.loadPrivateKey()
+            for user in users:
+                decrypted_username_bytes = cryptoUtils.decryptWithPrivateKey(private_key, user[3]) 
+                decrypted_username = decrypted_username_bytes.decode('utf-8')  # Assuming UTF-8 encoding
+                print(f"Decrypted username: {decrypted_username}")
+                if decrypted_username == username:
+                    return user
+
+            # If user is not found
+            print(f"User with username {username} not found.")
+            return None
+
         except sqlite3.Error as e:
             print("An error occurred while retrieving user data:", e)
             return None
@@ -142,53 +165,86 @@ class DB:
             if conn:
                 conn.close()
 
-    def findMembershipID(self, membership_id):
+
+    def findMembershipID(self, encrypted_membership_id):
         conn = None
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
-            query = "SELECT * FROM members WHERE membership_id = ?"
-            cursor.execute(query, (membership_id,))
-            
-            member = cursor.fetchone()
+            query = "SELECT * FROM members"
+            cursor.execute(query)
+            members = cursor.fetchall()
             cursor.close()
-            return member is not None
+
+            private_key = cryptoUtils.loadPrivateKey()  # Load private key for decryption
+            decrypted_membership_id = None
+
+            for member in members:
+                decrypted_membership_id = cryptoUtils.decryptWithPrivateKey(private_key, member[0])  # Assuming membership_id is in the 2nd column
+                if decrypted_membership_id.decode('utf-8') == encrypted_membership_id:
+                    return True  # Found matching membership ID
+
+            return False  # No matching membership ID found
+
         except sqlite3.Error as e:
             print("An error occurred while searching for membership ID:", e)
             return False
         finally:
             if conn:
                 conn.close()
+
     
     def findUserID(self, user_id, role):
         conn = None
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
-            query = "SELECT * FROM users WHERE role = ?"
-            cursor.execute(query, (role.value,))
-            
+            query = "SELECT * FROM users"
+            cursor.execute(query)
             users = cursor.fetchall()
             cursor.close()
-            return any(user[0] == user_id for user in users)
+
+            private_key = cryptoUtils.loadPrivateKey()  # Load private key for decryption
+            decrypted_role = None
+
+            for user in users:
+                decrypted_role = cryptoUtils.decryptWithPrivateKey(private_key, user[6])  # Assuming role is in the 7th column
+                if decrypted_role.decode('utf-8') == role.value:
+                    if user[0] == user_id:
+                        return True  # Found matching user_id and role
+
+            return False  # No matching user_id and role found
+
         except sqlite3.Error as e:
             print("An error occurred while searching for user ID:", e)
             return False
         finally:
             if conn:
                 conn.close()
+
     
     def findUsername(self, username):
         conn = None
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
-            query = "SELECT * FROM users WHERE username = ?"
-            cursor.execute(query, (username,))
+            query = "SELECT * FROM users"
+            cursor.execute(query)
             
-            user = cursor.fetchone()
+            users = cursor.fetchall()
             cursor.close()
-            return user is not None
+            
+            private_key = cryptoUtils.loadPrivateKey()
+            
+            for user in users:
+                decrypted_username_bytes = cryptoUtils.decryptWithPrivateKey(private_key, user[3])
+                decrypted_username = decrypted_username_bytes.decode('utf-8')  # Assuming UTF-8 encoding
+                
+                if decrypted_username == username:
+                    return True
+            
+            return False
+        
         except sqlite3.Error as e:
             print("An error occurred while searching for username:", e)
             return False
@@ -196,28 +252,39 @@ class DB:
             if conn:
                 conn.close()
 
+
     def getUsers(self, role=None):
         conn = None
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
             
-            if role is None:
-                query = "SELECT * FROM users WHERE role != 'superadmin'"
-                cursor.execute(query)
-            else:
-                query = "SELECT * FROM users WHERE role = ?"
-                cursor.execute(query, (role.value,))
+            query = "SELECT * FROM users"
+            cursor.execute(query)
             
             users = cursor.fetchall()
             cursor.close()
+            if role is not None:
+                private_key = cryptoUtils.loadPrivateKey()
+                userList = []
+                for user in users:
+                    encrypted_role = user[6]  # Assuming role is stored as encrypted bytes in user[6]
+                    decrypted_role_bytes = cryptoUtils.decryptWithPrivateKey(private_key, encrypted_role).decode('utf-8')
+                    if decrypted_role_bytes == role.value:
+                        userList.append(user)
+                
+                return userList
+            
             return users
+        
         except sqlite3.Error as e:
             print("An error occurred while retrieving users:", e)
             return None
+        
         finally:
             if conn:
                 conn.close()
+
     
     def getMembers(self):
         conn = None
@@ -263,7 +330,7 @@ class DB:
         try:
             conn = sqlite3.connect(self.databaseFile)
             hashed_password, salt = cryptoUtils.hashPassword(password)
-            query = "INSERT INTO users (first_name, last_name, username, password, registration_date, role, temp, salt) VALUES (?, ?, ?, ?, ?, ?, ?, ?)"
+            query = "INSERT INTO users (first_name, last_name, username, password_hash, registration_date, role, temp, salt) VALUES (?, ?, ?, ?, ?, ?, ?, ?)"
             parameters = (first_name, last_name, username, hashed_password, registration_date, role.value, temp, salt)
             cursor = conn.cursor()
 
@@ -283,35 +350,79 @@ class DB:
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
-            query = "DELETE FROM users WHERE id = ? AND role = ?"
-            cursor.execute(query, (user_id, role.value))
-            conn.commit()
-            if cursor.rowcount == 0:
-                raise ValueError(f"No user found with id {user_id} and role {role.value}")
-            cursor.close()
-            return "OK"
+            query = "SELECT * FROM users WHERE id = ?"
+            cursor.execute(query, (user_id,))
+            user = cursor.fetchone()
+            
+            if user is None:
+                raise ValueError(f"No user found with id {user_id}")
+            
+            private_key = cryptoUtils.loadPrivateKey()  # Load private key for decryption
+            decrypted_role = cryptoUtils.decryptWithPrivateKey(private_key, user[6])  # Assuming role is in the 7th column
+            
+            if decrypted_role.decode('utf-8') == role.value:
+                delete_query = "DELETE FROM users WHERE id = ? AND role = ?"
+                cursor.execute(delete_query, (user_id, role.value))
+                conn.commit()
+                
+                if cursor.rowcount == 0:
+                    raise ValueError(f"No user found with id {user_id} and role {role.value}")
+                
+                cursor.close()
+                return "OK"
+            else:
+                raise ValueError(f"Role mismatch for user with id {user_id}")
+        
         except sqlite3.Error as e:
             print(f"An error occurred while deleting the user: {e}")
             return "FAIL"
         except ValueError as ve:
             print(str(ve))
+            return "FAIL"
+        finally:
+            if conn:
+                conn.close()
+
 
     def deleteMember(self, membership_id):
         conn = None
         try:
             conn = sqlite3.connect(self.databaseFile)
             cursor = conn.cursor()
-            query = "DELETE FROM members WHERE membership_id = ?"
-            cursor.execute(query, (membership_id,))
-            conn.commit()
+
+            # Fetch all members
+            query = "SELECT * FROM members"
+            cursor.execute(query)
+            members = cursor.fetchall()
+
+            # Decrypt and find the member
+            private_key = cryptoUtils.loadPrivateKey()
+            found_member = False
+            for member in members:
+                decrypted_membership_id = cryptoUtils.decryptWithPrivateKey(private_key, member[0])  # Assuming membership_id is in the second column
+                if decrypted_membership_id.decode('utf-8') == membership_id:  # Decode bytes to string for comparison
+                    query = "DELETE FROM members WHERE membership_id = ?"
+                    cursor.execute(query, (member[0],))  # Pass the original encrypted membership_id to delete
+                    conn.commit()
+                    found_member = True
+                    break
+
             cursor.close()
-            return "OK"
+
+            if found_member:
+                return "OK"
+            else:
+                print(f"No member found with membership ID {membership_id}.")
+                return "NOT FOUND"
+
         except sqlite3.Error as e:
             print("An error occurred while deleting the member:", e)
-            return None
+            return "FAIL"
+
         finally:
             if conn:
                 conn.close()
+
 
     def updateUser(self, userId, firstName, lastName, username, role):
         conn = None
@@ -401,4 +512,7 @@ class DB:
         finally:
             if conn:
                 conn.close()
-
+dbPath = os.path.join(os.path.dirname(__file__), 'uniqueMeal.db')
+dataBase = DB(dbPath)
+list = dataBase.getUsers(None)
+print(list)
